PUB-507 - Moved to seperate file
